requirements.txt
numpy
paddleocr
pillow
streamlit
